extended the auth.user model
