+++ mpact/models.py
@@ -1,36 +1,19 @@
-from django.contrib.auth.models import (
-    AbstractBaseUser,
-    BaseUserManager,
-    PermissionsMixin,
-)
+from django.contrib.auth.models import User
+from django.core.exceptions import ValidationError
+from django.core.validators import RegexValidator
 from django.db import models
-from django.utils import timezone
 
 
-class UserManager(BaseUserManager):
-    def create_user(self, phone, password=None, **extra_fields):
-        user = self.model(phone=phone, **extra_fields)
-        user.set_password(password)
-        user.save(using=self._db)
-        return user
-
-    def create_superuser(self, phone, password, **extra_fields):
-        extra_fields.setdefault("is_staff", True)
-        extra_fields.setdefault("is_superuser", True)
-        return self.create_user(phone, password, **extra_fields)
+def validate_phone(username):
+    if username[0] != "+":
+        raise ValidationError("Phone number should start with '+'.")
+    return username
 
 
-class CustomUser(AbstractBaseUser, PermissionsMixin):
-
-    username = None
-    phone = models.CharField(max_length=20, unique=True)
-    is_staff = models.BooleanField(default=False)
-    is_superuser = models.BooleanField(default=False)
-    first_name = models.CharField(max_length=16)
-    last_name = models.CharField(max_length=16)
-    date_joined = models.DateTimeField(default=timezone.now)
-
-    USERNAME_FIELD = "phone"
-    REQUIRED_FIELDS = []
-
-    objects = UserManager()
+class Profile(models.Model):
+    user = models.OneToOneField(User, on_delete=models.CASCADE)
+    phone_regex = RegexValidator(
+        regex=r"^\+?1?\d{9,15}$",
+        message="Phone number is invalid.",
+    )
+    phone = models.CharField(max_length=20, validators=[phone_regex, validate_phone])
