+++ telegram_bot/settings.py
@@ -120,6 +120,3 @@ if os.environ.get("DEPLOY_ENV") == "dev":
 
 else:
     DATABASES = {"default": dj_database_url.config(conn_max_age=600, ssl_require=True)}
-
-
-AUTH_USER_MODEL = "mpact.CustomUser"
